Move TLP constants up to stix2 namespace
